Joblib update
This commit updates joblib to 0.12.0 and puts it as a submodule.
In the process of this update, minor alterations were made:

- the check_pickle argument to ``delayed``is deprecated in joblib,
  so calls to delayed don't call check_pickle.
- The default backend is set to ``spawn`` for all platforms, since
  there has been some issue with forking on unix
- The verbosity of ``joblib.Parallel`` is now always 0.
- unittests for sparse inputs now cast as 64-bit floats.

+++ mlens/parallel/tests/test_c_multi_sparse_feature_prop.py
@@ -15,7 +15,7 @@ SEED = 1324
 class TempClass(BaseEnsemble):
 
     def __init__(self):
-        super(TempClass, self).__init__()
+        super(TempClass, self).__init__(dtype=np.float64)
 
 
 def _shuffled(X, y, seed):
@@ -36,8 +36,8 @@ class OLSSparse(OLS):
         return super(OLSSparse, self).predict(X.toarray())
 
 
-X = np.random.rand(10, 50).astype(np.float32)
-y = np.arange(10).astype(np.float32)
+X = np.random.rand(10, 50)
+y = np.arange(10)
 
 first_prop = [1, 2, 3]
 n_first_prop = len(first_prop)
@@ -109,4 +109,4 @@ def test_shuffle():
 
     z = ens4.fit(X, y, return_preds=True)
 
-    np.testing.assert_array_equal(h.astype(np.float32), z)
+    np.testing.assert_array_equal(h, z)
